refactor: lift keycode_helper and oled_helper to shared polykybd level
Move keycode_helper.{c,h} and extract oled_helper.{c,h} to the shared
keyboards/handwired/polykybd/ directory so both split72 and corne42 build
from a single copy.

- polykybd/keycode_helper.{c,h}: replaces per-board copies; also fixes
  KC_KP_ASTERISK fall-through bug present in the split72 version
- polykybd/oled_helper.{c,h}: extracts num_to_u16_string, hex_to_u16_string,
  oled_status_screen, oled_render_logos, and oled_task_user from both
  status_oled.c files
- Both status_oled.h files re-export oled_helper.h so existing keymap.c
  includes continue to resolve the shared function declarations
- Per-board keycode_helper.{c,h} under each keymap/default/ deleted

Both handwired/polykybd/split72 and handwired/polykybd/corne42 build clean.
