+++ keyboards/handwired/polykybd/corne42/status_oled.h
@@ -2,8 +2,8 @@
 // SPDX-License-Identifier: GPL-2.0-or-later
 #pragma once
 
-void oled_render_logos(void);
+#include "../oled_helper.h"
 
+void oled_update_buffer(void);
 void oled_draw_kybd(void);
-
 void oled_draw_poly(void);
