+++ keyboards/handwired/polykybd/split72/keymaps/default/rules.mk
@@ -1,2 +1,1 @@
-SRC += keycode_helper.c
 ENCODER_MAP_ENABLE = yes
