+++ keyboards/handwired/polykybd/split72/status_oled.c
@@ -9,57 +9,16 @@
 #include "../base/com.h"
 #include "../base/disp_array.h"
 #include "../base/text_helper.h"
-//#include "../base/fonts/gfx_used_fonts.h"
 #include "../base/fonts/NotoSans_Regular_Base_11pt.h"
 #include "../base/fonts/NotoSans_Medium_Base_8pt.h"
-#include "../lang/named_glyphs.h" //should move to base or similar
+#include "../lang/named_glyphs.h"
+#include "../oled_helper.h"
 
 #include QMK_KEYBOARD_H
 #include "quantum.h"
 
 #include <stdint.h>
 #include <stdio.h>
-
-// Converts a number to a u16_string.
-// Global variables: (none - uses passed parameters only)
-void num_to_u16_string(char* buffer, uint8_t buffer_len, uint8_t value) {
-    if(value<10) {
-        snprintf((char*) buffer, buffer_len, "%d", value);
-        buffer[1] = 0;
-        buffer[2] = 0;
-        buffer[3] = 0;
-    } else if(value>99) {
-        snprintf((char*) buffer, buffer_len, "%d %d %d", value/100, (value/10)%10, value%10);
-        buffer[1] = 0;
-        buffer[3] = 0;
-        buffer[5] = 0;
-        buffer[6] = 0;
-        buffer[7] = 0;
-    } else {
-        snprintf((char*) buffer, buffer_len, "%d %d", value/10, value%10);
-        buffer[1] = 0;
-        buffer[3] = 0;
-        buffer[4] = 0;
-        buffer[5] = 0;
-    }
-}
-
-// Converts a number to a hex formated u16_string.
-// Global variables: (none - uses passed parameters only)
-void hex_to_u16_string(char* buffer, uint8_t buffer_len, uint8_t value) {
-    if(value<16) {
-        snprintf((char*) buffer, buffer_len, "%X", value);
-        buffer[1] = 0;
-        buffer[2] = 0;
-        buffer[3] = 0;
-    } else {
-        snprintf((char*) buffer, buffer_len, "%X %X", value/16, value%16);
-        buffer[1] = 0;
-        buffer[3] = 0;
-        buffer[4] = 0;
-        buffer[5] = 0;
-    }
-}
 
 //this should go away with some font refactoring!
 extern const GFXfont* ALL_FONTS [];
@@ -141,47 +100,6 @@ void oled_update_buffer(void) {
         kdisp_write_gfx_text(smallFont, 1, 84, 58, buffer);
     }
 }
-
-// Updates and displays status screen on OLED, or turns off display if status disabled.
-
-void oled_status_screen(void) {
-    const poly_sync_t* local_state = get_local_state();
-    if( (local_state->flags&STATUS_DISP_ON) == 0) {
-        oled_off();
-        return;
-    } else if( (local_state->flags&STATUS_DISP_ON) != 0) {
-        oled_on();
-    }
-
-    oled_update_buffer();
-    oled_clear();
-    oled_write_raw((char *)get_scratch_buffer(), get_scratch_buffer_size());
-}
-
-// Displays scrolling Polykybd logo on OLED with direction based on side (left/right).
-void oled_render_logos(void) {
-    if(is_left_side()) {
-        oled_draw_poly();
-        oled_scroll_right();
-    } else {
-        oled_draw_kybd();
-        oled_scroll_left();
-    }
-}
-
-// Main OLED task: displays logos during idle or status screen during normal operation.
-// Global variables: l_state
-bool oled_task_user(void) {
-    if((get_local_state()->flags&DISP_IDLE) != 0) {
-        oled_render_logos();
-    } else {
-        oled_scroll_off();
-        oled_status_screen();
-    }
-    return false;
-}
-
-
 
 void oled_draw_kybd(void) {
     static const char kybd_bitmap [] =
@@ -290,8 +208,6 @@ void oled_draw_kybd(void) {
     0x01, 0x03, 0x03, 0x0F, 0x0F, 0x0F, 0x03, 0x01, 0x00, 0x00,
     0x00, 0x00, 0x00, 0x00
     };
-
-
 
     oled_write_raw_P(kybd_bitmap, sizeof(kybd_bitmap));
 }
